Precomputed rectangle packing for totems 64 & 256
This is unfortunately a very big commit.

Includes logic to precompute all possible ways to build rectangles with
area up to a certain limit, using Knuth's Algorith X (using Dancing
Links, a.k.a DLX). Those are stored on disk. Commit includes precomputed
rectangles for area up to 32. This logic lives in 'dlx_solver.rs' and
'rect_inventory.rs'.

Includes logic to find all possible sums of rectangles that equal a
given totem bag (our question). This is analoguous to multi-dim subset
sum, with reuse, so this logic lives in 'subset_sum.rs'.

From a set of chosen rectangles, we try to do a rectangle packing, using
an "MCTS" approach (Monte Carlo Tree Search), where we pick the next
rectangle based on the one that scored best in a couple of sampled
simulations (best max or avg depth until "stuck"). This logic lives in
'rect_packing.rs'.
Rectangles are placed based on the Bottom-Left heuristic, where we pick
the lowest spot where it fits and left-align it.

To speed up the MCTS, we represent free spaces using "MaxRects" instead
of a grid of masks. We track free areas as maximal rectangles that get
split up when a piece is placed. This logic lives in 'max_rects.rs'.

With these combined, we get a decent (~10-20%) packing rate on the
hardest level (256 totems), compared to our greedy strategy that would
get <0.1%. We use 'automate.py' to retry every few minutes on the server
to hope and get an assigment where we find many perfect packs. After
running overnight, we got a score of 6012.50.