Add RectangleInventory and precompute_rects bin
Used to produce offline versions of rectangles that can be made up of
totem pieces. Added a RectPackingSolver skeleton to make use of it and
load the rectangles.

This does mean that running the binary won't work out of the box now
without generating precomputed rectangles beforehand, which can be done
with the following command:

```
cargo run --release --bin precompute_rects -- --area 32
```