Add subset sum (with repeats) iterator for rects
This allows to find combinations (with replacement) of rectangles that
have costs that sum up to a give totem bag. This is to be used in
conjunction with a rectangle packing algorithm to check if the set of
precomputed rectangles perfectly fits the target area.
Note that subset sum is a bit of a misnomer because of the repeats.