Add MaxRects structure
Instead of tracking "free" spots on a grid, use a MaxRects structure
from bin packing litterature to keep track of free spaces as rectangles.
MaxRects specifically expand the free spaces maximally in both
directions, which requires additional complexity in bookkeeping
(abstracted in this class), but makes finding a fit as easy as iterating
through the free spaces and picking any that fits (in our case, using a
Bottm-Left heuristic).